requirements.txt
selenium==4.0.0
pytest==7.0.1
allure-pytest==2.10.0
pytest-xdist==2.5.0
pyyaml==6.0
pytest-html==4.1.1